+++ js/main.js
@@ -1,79 +1,29 @@
-const HOUSE_TITLES = [
-  'Замок принцессы',
-  'Прекрасное жилище',
-  'Дом с видом на море',
-  'Дом с привидениями',
-];
-
-const MIN_HOUSE_COST = 5000;
-const MAX_HOUSE_COST = 1000000;
-
-const HOUSE_TYPES = [
-  'palace',
-  'flat',
-  'house',
-  'bungalow',
-];
-const MIN_HOUSE_ROOMS = 1;
-const MAX_HOUSE_ROOMS = 80;
-
-const MIN_HOUSE_GUESTS = 1;
-const MAX_HOUSE_GUESTS = 100;
-
-const TIME_POINTS = [
-  '12:00',
-  '13:00',
-  '14:00',
-];
-
-const HOUSE_DESCRIPTIONS = ['Чистые помещения, открытая веранда, вежливые соседи', 'Крыша немного подтекает, но это создаёт особый шарм' , 'Дом для приятного отдыха. Можно с детьми и животными', 'Развитая инфраструктура: рядом лес, парк и Пятёрочка'];
-const HOUSE_FEATURES = ['wifi', 'dishwasher', 'parking', 'washer', 'elevator', 'conditioner'];
-
-const HOUSE_PHOTOS = [
-  'http://o0.github.io/assets/images/tokyo/hotel1.jpg',
-  'http://o0.github.io/assets/images/tokyo/hotel2.jpg',
-  'http://o0.github.io/assets/images/tokyo/hotel3.jpg',
-]
-const MIN_COORDINATES_X = 35.65000;
-const MAX_COORDINATES_X = 35.70000;
-
-const MIN_COORDINATES_Y = 139.70000;
-const MAX_COORDINATES_Y = 139.80000;
-
-const COORDINATES_PRECISION = 5;
-
-// Получение числа с плавающей запятой
-const getRandomForSign = function (min, max, sign){
-  return (Math.random() * (max - min) + min).toFixed(sign);
-}
-
-// Получение рандомного целого числа
-const getRandomInteger = function (min, max){
-  min = Math.ceil(min);
-  max = Math.floor(max);
-  return Math.floor(Math.random() * (max - min + 1)) + min;
-}
-
-// create avatar
-const getRandomAvatar = () => {
-  return 'img/avatars/user0' + getRandomInteger(1, 8) +'.png';
-}
-
-// Получение рандомного элемента из массива
-const getRandomArrayElement = function (sourceArray){
-  let randomKey = Math.floor(Math.random() * sourceArray.length);
-  return sourceArray[randomKey];
-}
-
-// Получение массива случайной длины
-const getRandomlySlicedArray = (sourceArray) => {
-  let maxLength = Math.floor(Math.random() * (sourceArray.length));
-  let newArray = [];
-  for (let i = 0; i <= maxLength; i++){
-    newArray.push(sourceArray[i]);
-  }
-  return newArray;
-}
+import {
+  getRandomForSign,
+  getRandomInteger,
+  getRandomAvatar,
+  getRandomArrayElement,
+  getRandomlySlicedArray
+} from './util.js';
+import {
+  HOUSE_TITLES,
+  MIN_HOUSE_COST,
+  MAX_HOUSE_COST,
+  HOUSE_TYPES,
+  MIN_HOUSE_ROOMS,
+  MAX_HOUSE_ROOMS,
+  MIN_HOUSE_GUESTS,
+  MAX_HOUSE_GUESTS,
+  TIME_POINTS,
+  HOUSE_DESCRIPTIONS,
+  HOUSE_FEATURES,
+  HOUSE_PHOTOS,
+  MIN_COORDINATES_X,
+  MAX_COORDINATES_X,
+  MIN_COORDINATES_Y,
+  MAX_COORDINATES_Y,
+  COORDINATES_PRECISION
+} from './data.js';
 
 const getAd = () => {
   const ad = {
@@ -104,4 +54,4 @@ const getAd = () => {
 const getArray = (length = 0, cb) => {
   return Array.from({ length: length }, cb);
 }
-console.log(getArray(10, getAd));
+getArray(10, getAd);
