+++ package.json
@@ -1,4 +1,5 @@
 {
+  "type": "module",
   "name": "keksobooking",
   "version": "22.0.0",
   "private": true,
